requirements.txt
llama-index
openai
pypdf
python-dotenv
langchain_openai
langchain_core
streamlit
langchain_community
langserve
fastapi
uvicorn
sse-starlette
bs4
faiss-cpu
groq
langchain-groq
beautifulsoup4
